refractor(ClickHouseDriver): reorganize codes
