--- Setup.hs.save0
@@ -1,2 +1,0 @@
-import Distribution.Simple
-main = defaultMain
