+++ clickhouse-haskell.cabal
@@ -27,9 +27,10 @@ source-repository head
 
 library
   exposed-modules:
-      ClickHouseDriver.HttpConnection
-      ClickHouseDriver.Query
       ClickHouseDriver
+      ClickHouseDriver.Helpers
+      ClickHouseDriver.Query
+      ClickHouseDriver.Types
   other-modules:
       Paths_clickhouse_haskell
   hs-source-dirs:
